5288 Split out class QueryWebSocketClientAdapter
