+++ java/clients/src/main/java/sleeper/clients/query/QueryWebSocketClient.java
@@ -15,17 +15,10 @@
  */
 package sleeper.clients.query;
 
-import org.java_websocket.client.WebSocketClient;
-import org.java_websocket.handshake.ServerHandshake;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import software.amazon.awssdk.auth.credentials.AwsCredentials;
 import software.amazon.awssdk.auth.credentials.AwsCredentialsProvider;
-import software.amazon.awssdk.http.SdkHttpMethod;
-import software.amazon.awssdk.http.SdkHttpRequest;
-import software.amazon.awssdk.http.auth.aws.signer.AwsV4HttpSigner;
-import software.amazon.awssdk.http.auth.spi.internal.signer.DefaultSignRequest;
-import software.amazon.awssdk.http.auth.spi.signer.SignedRequest;
 
 import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.properties.table.TableProperties;
@@ -105,96 +98,12 @@ public class QueryWebSocketClient {
             QueryWebSocketMessageHandler messageHandler = new QueryWebSocketMessageHandler(tableProperties.getSchema());
             LOGGER.info("Obtaining AWS IAM credentials...");
             AwsCredentials credentials = credentialsProvider.resolveCredentials();
-            return new WebSocketQueryClient(region, serverUri, credentials, messageHandler);
+            return new QueryWebSocketClientAdapter(region, serverUri, credentials, messageHandler);
         };
     }
 
     public interface ClientProvider {
         Client createClient(InstanceProperties instanceProperties, TableProperties tableProperties);
-    }
-
-    private static class WebSocketQueryClient extends WebSocketClient implements Client {
-        private final String region;
-        private final URI serverUri;
-        private final AwsCredentials credentials;
-        private final QueryWebSocketMessageHandler messageHandler;
-        private Query query;
-
-        private WebSocketQueryClient(String region, URI serverUri, AwsCredentials credentials, QueryWebSocketMessageHandler messageHandler) {
-            super(serverUri);
-            this.region = region;
-            this.serverUri = serverUri;
-            this.credentials = credentials;
-            this.messageHandler = messageHandler;
-            messageHandler.setCloser(this::closeBlocking);
-        }
-
-        public CompletableFuture<List<String>> startQueryFuture(Query query) throws InterruptedException {
-            CompletableFuture<List<String>> future = new CompletableFuture<>();
-            messageHandler.setFuture(future);
-            this.query = query;
-            initialiseConnection();
-            return future;
-        }
-
-        private void initialiseConnection() throws InterruptedException {
-            setAwsIamAuthHeaders();
-            LOGGER.info("Connecting to WebSocket API at {}", serverUri);
-            connectBlocking();
-        }
-
-        private void setAwsIamAuthHeaders() {
-            LOGGER.debug("Creating auth signature with server URI: {}", serverUri);
-            AwsV4HttpSigner signer = AwsV4HttpSigner.create();
-            SignedRequest signed = signer.sign(DefaultSignRequest.builder(credentials)
-                    .putProperty(AwsV4HttpSigner.SERVICE_SIGNING_NAME, "execute-api")
-                    .putProperty(AwsV4HttpSigner.REGION_NAME, region)
-                    .request(SdkHttpRequest.builder()
-                            .uri(serverUri)
-                            .protocol("https")
-                            .method(SdkHttpMethod.GET)
-                            .build())
-                    .build());
-            LOGGER.debug("Setting auth headers...");
-            signed.request().forEachHeader((header, values) -> {
-                for (String value : values) {
-                    addHeader(header, value);
-                }
-            });
-        }
-
-        public boolean hasQueryFinished() {
-            return messageHandler.hasQueryFinished();
-        }
-
-        public long getTotalRecordsReturned() {
-            return messageHandler.getTotalRecordsReturned();
-        }
-
-        @Override
-        public void onOpen(ServerHandshake handshake) {
-            messageHandler.onOpen(query, this::send);
-        }
-
-        @Override
-        public void onMessage(String json) {
-            messageHandler.onMessage(json);
-        }
-
-        @Override
-        public void onClose(int code, String reason, boolean remote) {
-            messageHandler.onClose(reason);
-        }
-
-        @Override
-        public void onError(Exception error) {
-            messageHandler.onError(error);
-        }
-
-        @Override
-        public List<String> getResults(String queryId) {
-            return messageHandler.getResults(queryId);
-        }
     }
 
     public interface ClientCloser {
